requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
boto3==1.9.108
botocore==1.12.134
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
dj-database-url==0.5.0
Django==2.2.5
django-appconf==1.0.2
django-ckeditor==5.6.1
django-compressor==2.2
django-cors-headers==2.5.2
django-easy-select2==1.5.5
django-filter==2.1.0
django-graphql-jwt==0.2.0
django-hamlpy==1.2
django-imagekit==4.0.2
django-js-asset==1.1.0
django-sass-processor==0.7.2
django-storages==1.7.1
docutils==0.14
future==0.17.1
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
jmespath==0.9.4
libsass==0.17.0
oauth2client==4.1.3
oauthlib==3.0.1
pilkit==2.0
Pillow==5.4.1
promise==2.2.1
psycopg2==2.7.7
psycopg2-binary==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.1
python-telegram-bot==11.1.0
pytz==2018.9
rcssmin==1.0.6
regex==2019.1.24
requests==2.21.0
requests-oauthlib==1.2.0
rjsmin==1.0.12
rsa==4.0
Rx==1.6.1
s3transfer==0.2.0
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.2
whitenoise==4.1.2
